// backend/generated/prisma-chat/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const ReportType = {
  LOST: 'LOST',
  FOUND: 'FOUND',
  LAST_SEEN: 'LAST_SEEN'
} as const

export type ReportType = (typeof ReportType)[keyof typeof ReportType]


export const ReportStatus = {
  ACTIVE: 'ACTIVE',
  RECOVERED: 'RECOVERED'
} as const

export type ReportStatus = (typeof ReportStatus)[keyof typeof ReportStatus]


export const ChatRequestStatus = {
  PENDING: 'PENDING',
  ACCEPTED: 'ACCEPTED',
  DECLINED: 'DECLINED'
} as const

export type ChatRequestStatus = (typeof ChatRequestStatus)[keyof typeof ChatRequestStatus]


export const VerificationPurpose = {
  REGISTER: 'REGISTER',
  RESET_PASSWORD: 'RESET_PASSWORD'
} as const

export type VerificationPurpose = (typeof VerificationPurpose)[keyof typeof VerificationPurpose]